use rake and standard conventions for building/cleaning
Bundler ships with build, install, and release tasks
`rake build` will build the gem and drop it into pkg/

Rake ships with built-in clean and clobber tasks.
`rake clobber` will remove the pkg directory

There is no need for a custom build script. The clean and clobber tasks
are standard convention for rake. The build, install, and release tasks
are standard convention for bundler. We should use them.