Remove Axe::Page indirection
Axe::Page was *intended* to be an adapter that only exposed the bare
minimum of the webdriver/capybara api that was necessary within
axe-matchers. However, now that we are exposing hooks that will be
passing around the wrapped page object, we ought to be passing around a
fully wrapped object that has the full api delegated as expected. This
makes Axe::Page unnecessary